requirements.txt
psycopg2 >= 2.7.4
Django >= 2.0.4
pip >= 10.0.1
pipenv >= 11.10.0
django-extensions == 2.0.7
djangorestframework == 3.8.2
pillow >=5.1.0
django-cors-headers>=2.2.0
django-cleanup
model_mommy==1.5.1
pytest-django
pytest>=3.3.2
django-debug-toolbar
django-filter